+++ bidict/_orderedbidict.py
@@ -60,7 +60,7 @@ class OrderedBidict(OrderedBidictBase[KT, VT], MutableBidict[KT, VT]):
         """
         if not self:
             raise KeyError('mapping is empty')
-        key = next((reversed if last else iter)(self))
+        key = next((reversed if last else iter)(self))  # type: ignore
         val = self._pop(key)
         return key, val